Change text of ouroboros badges to better reflect level ranges
Based on testing, it seems like all the Ouroboros ranges start at 5n+1 and end at 5n. That said, this will likely never be 100% accurate as there are some that work for a completely different level range (there's a level 25 villain arc that awards the 1-15 badges for instance).

Also, standardized some text to make the descriptions uniform.