Update former-servant-of-recluse.ts
Updated this badge to use the correct icon (which is the icon for the 'Recluse's Betrayer' accolade badge.

Another way of fixing this (which I don't think I have the proper permissions for) is to reject this proposed change, copy 'recluses-betrayer.png' from coh-content-db-homecoming/images/badges/accolade to coh-content-db-homecoming/images/badges/accomplishment and use it to replace the existing 'former-servant-of-recluse.png' icon. The 'former-servant-of-recluse.png' icon is not referenced anywhere else (at least not as revealed by my check of the badge .ts files) so I think it is safe to overwrite that png file with the proper 'recluses-betrayer.png'.