Update patroller.ts
This is a strange one. It has a Praetorian variant that has the same name as the hero variant. However, according to the wiki, if a character is of Praetorian origin, then regardless of their alignment, they earn the Patroller badge. Therefore I added the Praetorian ID, the Praetorian name variant, and the Praetorian badge text variant.