Add files via upload
Adding images for new badges. The char-creation.png file is a more generically named version of the vip.png, which is now used by three different badges. vip.png can be deleted, as the other badges which use that image have been edited to now refer to char-creation.png instead.